requirements.txt
python-levenshtein
requests
tqdm